[firebase_database] Support v2 android embedder. (#287)
