+++ packages/firebase_database/example/pubspec.yaml
@@ -6,12 +6,17 @@ dependencies:
     sdk: flutter
   firebase_database:
     path: ../
-  firebase_core: ^0.4.0
+  firebase_core: ^0.4.1
 
 dev_dependencies:
   flutter_driver:
     sdk: flutter
   test: any
+  e2e: ^0.2.0
 
 flutter:
   uses-material-design: true
+
+environment:
+  sdk: ">=2.0.0-dev.28.0 <3.0.0"
+  flutter: ">=1.9.1+hotfix.2 <2.0.0"
