+++ packages/firebase_database/example/test_driver/firebase_database_e2e.dart
@@ -1,17 +1,14 @@
-import 'dart:async';
-import 'package:flutter_driver/driver_extension.dart';
+import 'package:e2e/e2e.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:firebase_database/firebase_database.dart';
 
 void main() {
-  final Completer<String> completer = Completer<String>();
-  enableFlutterDriverExtension(handler: (_) => completer.future);
-  tearDownAll(() => completer.complete(null));
+  E2EWidgetsFlutterBinding.ensureInitialized();
 
   group('$FirebaseDatabase', () {
     final FirebaseDatabase database = FirebaseDatabase.instance;
 
-    test('runTransaction', () async {
+    testWidgets('runTransaction', (WidgetTester tester) async {
       final DatabaseReference ref = database.reference().child('counter');
       final DataSnapshot snapshot = await ref.once();
       final int value = snapshot.value ?? 0;
