--- packages/firebase_database/example/test_driver/firebase_database_test.dart
@@ -1,10 +1,0 @@
-import 'package:flutter_driver/flutter_driver.dart';
-import 'package:test/test.dart';
-
-void main() {
-  test('FirebaseDatabase driver test', () async {
-    final FlutterDriver driver = await FlutterDriver.connect();
-    await driver.requestData(null, timeout: const Duration(minutes: 1));
-    driver.close();
-  });
-}
