+++ packages/firebase_database/pubspec.yaml
@@ -3,7 +3,7 @@ description: Flutter plugin for Firebase Database, a cloud-hosted NoSQL database
   with realtime data syncing across Android and iOS clients, and offline access.
 author: Flutter Team <flutter-dev@googlegroups.com>
 homepage: https://github.com/FirebaseExtended/flutterfire/tree/master/packages/firebase_database
-version: 3.0.9
+version: 3.1.0
 
 flutter:
   plugin:
@@ -23,7 +23,8 @@ dev_dependencies:
     sdk: flutter
   flutter_driver:
     sdk: flutter
+  e2e: ^0.2.1
 
 environment:
   sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=1.5.0 <2.0.0"
+  flutter: ">=1.9.1+hotfix.2 <2.0.0"
